+++ Source/DafnyLanguageServer/Workspace/ChangeProcessors/IRelocator.cs
@@ -2,6 +2,7 @@
 using Microsoft.Dafny.LanguageServer.Language.Symbols;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using System.Threading;
+using Microsoft.Dafny.LanguageServer.Workspace.Notifications;
 
 namespace Microsoft.Dafny.LanguageServer.Workspace.ChangeProcessors {
   /// <summary>
@@ -30,5 +31,15 @@ namespace Microsoft.Dafny.LanguageServer.Workspace.ChangeProcessors {
     /// <exception cref="System.OperationCanceledException">Thrown when the cancellation was requested before completion.</exception>
     /// <exception cref="System.ObjectDisposedException">Thrown if the cancellation token was disposed before the completion.</exception>
     IReadOnlyList<Diagnostic> RelocateDiagnostics(IReadOnlyList<Diagnostic> originalDiagnostics, DidChangeTextDocumentParams changes, CancellationToken cancellationToken);
+
+    /// <summary>
+    /// Relocate verification trees from a document to the next document, prior to verifying the new document.
+    /// That way, we can instantly publish old verifications trees for the positions in the new document
+    /// </summary>
+    /// <param name="oldVerificationTree">The verification tree that should be relocated, including its children</param>
+    /// <param name="documentChange">The applied changes to the text document that should be used for the relocation.</param>
+    /// <param name="cancellationToken">A token to stop the relocation prior completion.</param>
+    /// <returns></returns>
+    VerificationTree RelocateVerificationTree(VerificationTree oldVerificationTree, DidChangeTextDocumentParams documentChange, CancellationToken cancellationToken);
   }
 }