+++ Source/DafnyLanguageServer/Workspace/LanguageServerExtensions.cs
@@ -46,7 +46,8 @@ namespace Microsoft.Dafny.LanguageServer.Workspace {
         services.GetRequiredService<ISymbolTableFactory>(),
         services.GetRequiredService<IGhostStateDiagnosticCollector>(),
         services.GetRequiredService<ICompilationStatusNotificationPublisher>(),
-        services.GetRequiredService<ILoggerFactory>()
+        services.GetRequiredService<ILoggerFactory>(),
+        services.GetRequiredService<IDiagnosticPublisher>()
       );
     }
   }